backend/requirements.txt
Flask
mysql-connector-python
flask-cors
Pillow
matplotlib
numpy
opencv-python